refactor(md): refactor managed resource history + fix memoization
